backend/requirement.txt
fastapi
uvicorn
python-multipart
PyPDF2
python-docx
python-pptx
openai
redis
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
httpx